[cesium] - fixed a deadlock with relay on db closing (#673)
* [cesium] - fixed a deadlock with relay on db closing

* [cesium] - added close guards on delete and GC methods in unary

+++ cesium/internal/unary/delete.go
@@ -20,12 +20,18 @@ import (
 // Delete deletes the specified time range from the database. Note that the start of the
 // time range is inclusive whereas the end is note.
 func (db *DB) Delete(ctx context.Context, tr telem.TimeRange) error {
+	if db.closed.Load() {
+		return errDBClosed
+	}
 	return db.wrapError(db.delete(ctx, tr))
 }
 
 // GarbageCollect removes unused telemetry data in the unaryDB. It is NOT safe to call
 // concurrently with other GarbageCollect methods.
 func (db *DB) GarbageCollect(ctx context.Context) error {
+	if db.closed.Load() {
+		return errDBClosed
+	}
 	db.entityCount.add(1)
 	defer db.entityCount.add(-1)
 	return db.wrapError(db.Domain.GarbageCollect(ctx))
